move the getSvgCode to the main tools directory for public use
