+++ src/utils/tools/getSvgCode.ts
