+++ src/utils/tools/index.ts
@@ -7,5 +7,5 @@ export { default as zipFolder } from './zipFolder'
 export { default as getThemeDir } from './getThemeDir'
 export { default as getEncodedLink } from './updatingIcons/getEncodedLink'
 export { default as isNewIcon } from './updatingIcons/isNewIcon'
-export { getOutlined, getFilled } from './updatingIcons/getSvgCode'
+export { getOutlined, getFilled } from './getSvgCode'
 export { default as prepareUpdatedIcon } from './updatingIcons/prepareTheUpdate'
